+++ event/main_listener.php
@@ -1,8 +1,7 @@
 <?php
 /**
- * This file is part of the PM Search extension for phpBB 3.1/3.2.
+ * This file is part of the PM Search extension for phpBB.
  * @package bbw/pmsearch
- * @copyright (c) Stanislav Atanasov
  * @license GNU General Public License, version 2 [GPL-2.0](https://opensource.org/license/gpl-2-0)
  */
 namespace bbw\pmsearch\event;
@@ -17,10 +16,10 @@ class main_listener implements EventSubscriberInterface
 	static public function getSubscribedEvents()
 	{
 		return array(
-			'core.permissions'	       => 'acl_perms_add',
-			'core.submit_pm_after'	       => 'pm_search_main',
-			'core.delete_pm_before'	=> 'pm_delete_index',
-			'core.memberlist_view_profile'	       => 'pm_search_with_user',
+			'core.permissions'				=> 'acl_perms_add',
+			'core.submit_pm_after'			=> 'pm_search_main',
+			'core.delete_pm_before'			=> 'pm_delete_index',
+			'core.memberlist_view_profile'	=> 'pm_search_with_user',
 		);
 	}
 
@@ -42,9 +41,14 @@ class main_listener implements EventSubscriberInterface
 	 * @param \phpbb\template|\phpbb\template\template           $template  Template object
 	 * @internal param \phpbb\content_visibility $content_visibility Content visibility object
 	 */
-	public function __construct(\phpbb\config\config $config, \phpbb\db\driver\driver_interface $db,
+	public function __construct(
+		\phpbb\config\config $config,
+		\phpbb\db\driver\driver_interface $db,
 		\phpbb\template\template $template,
-		\bbw\pmsearch\helper $search_helper)
+		\bbw\pmsearch\helper $search_helper,
+		\bbw\pmsearch\fulltext_search $fulltext_search,
+		\bbw\pmsearch\search $search,
+	)
 	{
 		$this->config = $config;
 		$this->db = $db;
@@ -53,6 +57,7 @@ class main_listener implements EventSubscriberInterface
 
 		$error = false;
 		$search_types = $this->search_helper->get_search_types();
+
 		if ($this->search_helper->init_search($search_types[0], $this->fulltext_search, $error))
 		{
 			trigger_error($error, E_USER_WARNING);