+++ helper.php
@@ -1,17 +1,24 @@
 <?php
 /**
- * This file is part of the PM Search extension for phpBB 3.1/3.2.
+ * This file is part of the PM Search extension for phpBB.
  * @package bbw/pmsearch
- * @copyright (c) Anavaro
  * @license GNU General Public License, version 2 [GPL-2.0](https://opensource.org/license/gpl-2-0)
  */
 namespace bbw\pmsearch;
 
 class helper
 {
-	public function __construct(\phpbb\extension\manager $ext_manager, \phpbb\auth\auth $auth, \phpbb\config\config $config,
-		\phpbb\db\driver\driver_interface $db, \phpbb\user $user, \phpbb\language\language $language, \phpbb\event\dispatcher $dispatcher,
-		$phpbb_root_path, $phpEx)
+	public function __construct(
+		\phpbb\extension\manager $ext_manager,
+		\phpbb\auth\auth $auth,
+		\phpbb\config\config $config,
+		\phpbb\db\driver\driver_interface $db,
+		\phpbb\user $user,
+		\phpbb\language\language $language,
+		\phpbb\event\dispatcher $dispatcher,
+		$phpbb_root_path,
+		$phpEx,
+	)
 	{
 		$this->ext_manager = $ext_manager;
 		$this->auth = $auth;