+++ language/en/permissions_pmsearch.php
@@ -34,20 +34,7 @@ if (empty($lang) || !is_array($lang))
 // ’ » “ ” …
 //
 
-$lang = array_merge($lang, array(
-	// UCP general langauge
-	'UCP_PMSEARCH_TITLE'	=> 'Search',
-
-	'PMSEARCH_TITLE'	=> 'Search in PMs',
-	'PMSEARCH_KEYWORDS_EXPLAIN'	=>	'Place + in front of a word which must be found and - in front of a word which must not be found. Put a list of words separated by | into brackets if only one of the words must be found. Use * as a wildcard for partial matches.',
-	'SEARCH_ALL_TERMS'	=>	' Search for all terms or use query as entered',
-	'SEARCH_ANY_TERMS'	=>	'Search for any terms',
-	'NO_RESULTS_FOUND'	=> 'No results found.',
-	'SEARCH_PMS'	=> 'Search PMs',
-	'ACCESS_DENIED'	=> 'You have no authority to search in PMs',
-
-	// Added in version 1.0.1
-	'SEARCH_FOR_NICK'	=> 'Search with nick',
-	'SEARCH_WITH_USER'	=> 'Search',
-	'SEARCH_WITH_USER_LANG'	=> 'Show conversations with user',
-));
+$lang = array_merge($lang, [
+	// Shared language strings
+	'ACP_PMSEARCH'	=> 'PM Search',
+]);