+++ migrations/release_1_0_0.php
@@ -1,8 +1,7 @@
 <?php
 /**
- * This file is part of the PM Search extension for phpBB 3.1/3.2.
+ * This file is part of the PM Search extension for phpBB.
  * @package bbw/pmsearch
- * @copyright (c) Stanislav Atanasov
  * @license GNU General Public License, version 2 [GPL-2.0](https://opensource.org/license/gpl-2-0)
  */
 namespace bbw\pmsearch\migrations;