+++ ucp/pmsearch_module.php
@@ -1,15 +1,17 @@
 <?php
 /**
- * This file is part of the PM Search extension for phpBB 3.1/3.2.
+ * This file is part of the PM Search extension for phpBB.
  * @package bbw/pmsearch
- * @copyright (c) Stanislav Atanasov
  * @license GNU General Public License, version 2 [GPL-2.0](https://opensource.org/license/gpl-2-0)
  */
 namespace bbw\pmsearch\ucp;
 
+/**
+ */
 class ucp_pmsearch_module
 {
 	var $u_action;
+
 	private $search_helper;
 	private $config;
 	private $terms_ary = array(
@@ -17,11 +19,14 @@ class ucp_pmsearch_module
 		'any'	=> 2,
 		'nick'	=> 3,
 	);
-	function main($id, $mode)
+
+	function main( $id, $mode )
 	{
 		global $db, $user, $auth, $template, $request, $phpbb_container, $config;
+
 		$this->config = $config;
 		$this->search_helper = $phpbb_container->get('bbw.pmsearch.search.helper');
+
 		if (!$auth->acl_get('u_pmsearch'))
 		{
 			trigger_error('ACCESS_DENIED');
@@ -169,6 +174,8 @@ class ucp_pmsearch_module
 				$template->assign_vars(array(
 					'SEARCH_TEARM_TYPE' => $this->terms_ary[$terms]
 				));
+
+			break;
 		}
 	}
 }